+++ server/src/main/java/lsfusion/server/form/instance/filter/IsClassFilterInstance.java
@@ -30,7 +30,7 @@ public class IsClassFilterInstance<P extends PropertyInterface> extends Property
 
     public IsClassFilterInstance(DataInputStream inStream, FormInstance form) throws IOException {
         super(inStream, form);
-        isClass = form.getCustomClass(inStream.readInt());
+        isClass = form.getCustomClass(inStream.readLong());
     }
 
     public Where getWhere(ImMap<ObjectInstance, ? extends Expr> mapKeys, Modifier modifier, ReallyChanged reallyChanged, MSet<CalcProperty> mUsedProps) throws SQLException, SQLHandledException {
